feat: ✨ show画面に指板図を表示する
