+++ app/javascript/controllers/index.js
@@ -2,10 +2,13 @@
 // Run that command whenever you add a new controller or create them with
 // ./bin/rails generate stimulus controllerName
 
-import { application } from "./application";
+import { application } from "./application"
 
-import HelloController from "./hello_controller";
-application.register("hello", HelloController);
+import HelloController from "./hello_controller"
+application.register("hello", HelloController)
 
-import ScoreController from "./score_controller";
-application.register("score", ScoreController);
+import ScoreController from "./score_controller"
+application.register("score", ScoreController)
+
+import ScoreFormController from "./score_form_controller"
+application.register("score-form", ScoreFormController)
